requirements.txt
APScheduler==3.5.3
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
funcsigs==1.0.2
futures==3.2.0
idna==2.7
olefile==0.45.1
Pillow==5.3.0
PyPDF2==1.26.0
pytz==2018.7
reportlab==3.5.10
requests==2.20.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.24.1
